# Conflicts:
#	build-utils/last-built-plugin.ts
#	config/rspack.chartcuterie.config.ts
#	package.json
#	rspack.config.ts
#	yarn.lock

+++ src/sentry/api/invite_helper.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import dataclasses
 from logging import Logger
-from typing import Dict, Optional
 
 from django.http.request import HttpRequest
 from django.utils.crypto import constant_time_compare
@@ -10,14 +9,14 @@ from django.utils.crypto import constant_time_compare
 from sentry import audit_log, features
 from sentry.models.authidentity import AuthIdentity
 from sentry.models.authprovider import AuthProvider
-from sentry.models.user import User
-from sentry.models.useremail import UserEmail
-from sentry.services.hybrid_cloud.organization import (
+from sentry.organizations.services.organization import (
     RpcOrganizationMember,
     RpcUserInviteContext,
     organization_service,
 )
 from sentry.signals import member_joined
+from sentry.users.models.user import User
+from sentry.users.models.useremail import UserEmail
 from sentry.utils import metrics
 from sentry.utils.audit import create_audit_entry
 
@@ -40,9 +39,9 @@ def remove_invite_details_from_session(request: HttpRequest) -> None:
 
 @dataclasses.dataclass
 class InviteDetails:
-    invite_token: Optional[str]
-    invite_member_id: Optional[int]
-    invite_organization_id: Optional[int]
+    invite_token: str | None
+    invite_member_id: int | None
+    invite_organization_id: int | None
 
 
 def get_invite_details(request: HttpRequest) -> InviteDetails:
@@ -221,10 +220,11 @@ class ApiInviteHelper:
 
         if self.member_already_exists:
             self.handle_member_already_exists()
-            organization_service.delete_organization_member(
-                organization_member_id=self.invite_context.invite_organization_member_id,
-                organization_id=self.invite_context.organization.id,
-            )
+            if self.invite_context.invite_organization_member_id is not None:
+                organization_service.delete_organization_member(
+                    organization_member_id=self.invite_context.invite_organization_member_id,
+                    organization_id=self.invite_context.organization.id,
+                )
             return None
 
         try:
@@ -289,7 +289,7 @@ class ApiInviteHelper:
         )
         return not primary_email_is_verified
 
-    def get_onboarding_steps(self) -> Dict[str, bool]:
+    def get_onboarding_steps(self) -> dict[str, bool]:
         return {
             "needs2fa": self._needs_2fa(),
             "needsEmailVerification": self._needs_email_verification(),
